Fix compilation issues when compiling against Linux versions greater than 4.11
In Linux 4.11.0, the vma argument was removed from the
vm_operations_struct::fault function.

In Linux 4.11.9, the netdevice destructor function was renamed to
priv_destructor and a needs_free_netdev member was added to the structure,
which should be set to true if the kernel module does not free its own struct
net_device upon exit.